* new MRC: many more bugfixes
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@83 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/src/org/xtreemfs/new_mrc/utils/Converter.java
@@ -65,7 +65,7 @@ public class Converter {
         Map<String, Object> aclMap = new HashMap<String, Object>();
         while (acl.hasNext()) {
             ACLEntry next = acl.next();
-            aclMap.put(next.getEntity(), next.getRights());
+            aclMap.put(next.getEntity(), (long) next.getRights());
         }
         
         return aclMap;
